# docs/swagger.yaml
basePath: /api
definitions:
  api.DefaultResponse:
    properties:
      code:
        description: code 为1表示正常 0表示业务请求错误
        type: integer
      msg:
        description: 错误提示信息
        type: string
    type: object
  api.SuccessResponse:
    properties:
      code:
        description: code 为1表示正常 0表示业务请求错误
        type: integer
      data:
        description: 接口返回的业务数据
        type: object
      msg:
        description: 错误提示信息
        type: string
    type: object
  models.UserLoginRes:
    properties:
      jtoken:
        description: Jtoken 验证字符串
        type: string
      retoken:
        description: retoken 刷新token
        type: string
    type: object
  models.UserReq:
    properties:
      age:
        description: 年龄
        type: integer
      sex:
        description: 性别
        type: integer
      username:
        description: 用户名
        type: string
    required:
    - age
    - sex
    - username
    type: object
host: localhost:20011
info:
  contact:
    name: gphper
    url: https://eGame-demo-back-office-api
  description: GinAdmin 示例项目
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: GinAdmin Api
  version: "1.0"
paths:
  /example/index:
    post:
      consumes:
      - multipart/form-data
      operationId: "1"
      parameters:
      - description: token
        in: header
        name: authorization
        required: true
        type: string
      - description: 年龄
        in: formData
        name: age
        required: true
        type: integer
      - description: 性别
        in: formData
        name: sex
        required: true
        type: integer
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserReq'
              type: object
        default:
          description: ""
          schema:
            $ref: '#/definitions/api.DefaultResponse'
      summary: 展示用户信息
      tags:
      - 示例
  /user//refresh:
    post:
      consumes:
      - multipart/form-data
      operationId: "4"
      parameters:
      - in: formData
        name: retoken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":1,"msg":"success","data":{"jtoken":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJFeHAiOiIyMDIxLTEyLTI2VDE5OjI1OjI4Ljg0OTIzNzUrMDg6MDAiLCJOYW1lIjoiZ3BocGVyIiwiVWlkIjo0fQ==.ab81bb7134978afe976df55b45789aefd10f6c3edb969bae283c32c080083b89"}}
          schema:
            type: json
        default:
          description: ""
          schema:
            $ref: '#/definitions/api.DefaultResponse'
      summary: 刷新jtoken
      tags:
      - 用户
  /user/login:
    post:
      consumes:
      - multipart/form-data
      operationId: "3"
      parameters:
      - in: formData
        name: email
        required: true
        type: string
      - in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserLoginRes'
        default:
          description: ""
          schema:
            $ref: '#/definitions/api.DefaultResponse'
      summary: 用户登录
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - multipart/form-data
      operationId: "2"
      parameters:
      - in: formData
        name: confirm_password
        required: true
        type: string
      - in: formData
        name: email
        required: true
        type: string
      - in: formData
        name: nickname
        required: true
        type: string
      - in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/api.DefaultResponse'
      summary: 用户注册
      tags:
      - 用户
swagger: "2.0"